+++ client/public/index.html
@@ -8,7 +8,7 @@
     <meta
       name="description"
       content="Web site created using create-react-app"
-    />
+     
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     
     <!--
@@ -16,6 +16,7 @@
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/ionicons/2.0.1/css/ionicons.min.css"/>
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
