+++ client/src/components/footer.js
@@ -13,7 +13,7 @@ function Footer() {
                 </div>
                 <ul className="list-inline">
                     <li className="list-inline-item"><a href="/">Home</a></li>
-                    <li className="list-inline-item"><a href="/">Services</a></li>
+                    <li className="list-inline-item"><a href="/">Register</a></li>
                     <li className="list-inline-item"><a href="/">About</a></li>
                     <li className="list-inline-item"><a href="/">Terms</a></li>
                 </ul>
