+++ client/src/index.css
@@ -1,4 +1,4 @@
-/* styles for components starts here
+ /* styles for components starts here
 
 :root {
   --primary:RGB(143, 62, 63);
@@ -14,7 +14,7 @@ main {
 	max-width: 90%;
 	padding: 50px 30px;
 }
-body {
+/* body {
 	margin: 0;
 	padding: 0;
 	font-weight: 600;
@@ -26,6 +26,15 @@ body {
 	font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
   
 }
+.heading {
+	background: white;
+	display: flex;
+	justify-content: space-between;
+	padding: 1rem;
+	text-align: center;
+	color: var(--primary);
+}
+*/
 a
 {
     text-decoration: none;
@@ -59,25 +68,40 @@ p {
 	margin-top: 0;
 }
 
-
-
-p {
-	font-size: 1.1rem;
-	margin-top: 0;
-}
-header {
-	background: white;
+.center {
 	display: flex;
-	justify-content: space-between;
-	padding: 1rem;
-	text-align: center;
+	justify-content: center;
+	
 }
+
+ .page-header {
+	font-family: 'Birthstone Bounce', cursive;
+	font-size: 3.5em;
+	font-weight: 500;
+	margin: 5px 0;
+	line-height: 1.25;
+	padding-bottom: 5rem;
+	
+	} 
+	
 
 span[role="img"] {
 	cursor: default;
 }
 
-/*nav section */
+
+
+/* map section */
+
+.map-container {
+
+margin-top: 10%;
+}
+
+
+/*map section styling ends */
+
+/*nav section starts  */
 
 ol li {
 	list-style: inside;
@@ -94,11 +118,20 @@ nav li {
 	color: var(--tertiary);
 	
 }
+.mx-2, .navActive {
+	color:#01474ef3;
+	font-weight: 600;
+	line-height: 1.5;
+	font-size: 18px;
+	letter-spacing: .1em;
+	font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+	
+}
 
 /* end nav section */
 
 img {
-	max-width: 20%;
+	max-width: 15%;
 	float: left;
 	
 }
@@ -112,71 +145,27 @@ img {
 .flex-row {
 	display: flex;
 	justify-content: space-evenly;
+	align-self: flex-end;
 
 }
 
 .space-between {
 	justify-content: space-between;
+	
 }
 
 .flex-end {
 	justify-content: flex-end;
 }
 
-.mx-1 {
-	margin-right: var(--spacing-one);
-	margin-left: var(--spacing-one);
-}
-
-.mx-2 {
-	margin-right: var(--spacing-two);
-	margin-left: var(--spacing-two);
-}
-
-.my-1 {
-	margin-top: var(--spacing-one);
-	margin-bottom: var(--spacing-one);
-}
-
-.my-2 {
-	margin-top: var(--spacing-two);
-	margin-bottom: var(--spacing-two);
-}
 
 
-.my-5 {
-	margin-top: var(--spacing-five);
-	margin-bottom: var(--spacing-five);
-}
-
-.px-1 {
-	padding-right: var(--spacing-one);
-	padding-left: var(--spacing-one);
-}
-
-.px-2 {
-	padding-right: var(--spacing-two);
-	padding-left: var(--spacing-two);
-}
-
-.py-1 {
-	padding-top: var(--spacing-one);
-	padding-bottom: var(--spacing-one);
-}
-
-.py-2 {
-	padding-top: var(--spacing-two);
-	padding-bottom: var(--spacing-two);
-}
-
-
-
-/* footer section */
+/* footer section starts  */
 
 .footer {
-	padding:40px 0;
-	background-color:#ffffff;
-	color: RGB(143, 62, 63);
+	margin:100px 0 0 0;
+	background-color: whitesmoke;
+	color: #01474ef3;
   }
   
   .footer ul {
@@ -229,27 +218,13 @@ img {
 	margin-top:15px;
 	text-align:center;
 	font-size:13px;
-	color:RGB(143, 62, 63);;
+	color:#01474ef3;
 	margin-bottom:0;
   }
 
   /* end footer section */
 
 
-.center {
-	display: flex;
-	justify-content: center;
-}
-
-.page-header {
-	font-family: 'Birthstone Bounce', cursive;
-	font-size: 2.5em;
-	font-weight: 500;
-	margin: 5px 0;
-	line-height: 1.25;
-	padding-bottom: 2rem;
-	
-}
